ecommerce_web_proj/requirements.txt
Django==4.2.7
gunicorn==21.2.0
psycopg2-binary==2.9.6
django-cors-headers==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
Pillow==10.1.0
stripe==7.3.0